backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.200
transformers>=4.30.0
sentencepiece>=0.1.99
pytesseract>=0.3.10
aiofiles>=23.0.0
python-dotenv>=1.0.0
scipy>=1.11.0
scikit-image>=0.21.0